requirements.txt
langchain>=0.2.0
langchain-nvidia-ai-endpoints>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
torch>=2.0.0
transformers>=4.40.0
pandas>=2.0.0
loguru>=0.7.0
